+++ NWN.Anvil/src/main/API/Events/Game/StoreEvents/StoreEvents.OnClose.cs
@@ -7,27 +7,8 @@ namespace Anvil.API.Events
   /// <summary>
   /// Events for Merchant/Store objects.
   /// </summary>
-  public static class StoreEvents
+  public static partial class StoreEvents
   {
-    [GameEvent(EventScriptType.StoreOnOpen)]
-    public sealed class OnOpen : IEvent
-    {
-      /// <summary>
-      /// Gets the <see cref="NwStore"/> being open.
-      /// </summary>
-      public NwStore Store { get; } = NWScript.OBJECT_SELF.ToNwObject<NwStore>();
-
-      /// <summary>
-      /// Gets the <see cref="NwPlayer"/> that last opened this store.
-      /// </summary>
-      public NwPlayer Player { get; } = NWScript.GetLastOpenedBy().ToNwPlayer();
-
-      NwObject IEvent.Context
-      {
-        get => Store;
-      }
-    }
-
     [GameEvent(EventScriptType.StoreOnClose)]
     public sealed class OnClose : IEvent
     {
@@ -53,13 +34,6 @@ namespace Anvil.API
 {
   public sealed partial class NwStore
   {
-    /// <inheritdoc cref="StoreEvents.OnOpen"/>
-    public event Action<StoreEvents.OnOpen> OnOpen
-    {
-      add => EventService.Subscribe<StoreEvents.OnOpen, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
-      remove => EventService.Unsubscribe<StoreEvents.OnOpen, GameEventFactory>(this, value);
-    }
-
     /// <inheritdoc cref="StoreEvents.OnClose"/>
     public event Action<StoreEvents.OnClose> OnClose
     {
